Complete_Auth_Jwt/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-dotenv==1.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0
